requirements.txt
numpy==2.0.2
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.13.1
dask==2024.8.0
dask-expr==1.1.10
cloudpickle==3.1.0
fsspec==2024.12.0
pyarrow==18.1.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
tqdm==4.67.1
joblib==1.4.2
threadpoolctl==3.5.0
pytest==8.3.4
pytest-progress==1.3.0
execnet==2.1.1 
pytest-cache==1.0
coverage==7.6.10 
pytest-cov==6.0.0